now grabs both probes, doesn't duplicate data
Previously, there was one directory per event, and it only held data for the probe that saw the event. Now, we'll have the event listing locally, and data will be organized by day (since we grabbed the whole day of data anyway). We grab the data for both probes for each day where either sees an event.