we have the complete data now. let's look for events.
also met with John.

A concern: E dot B = 0 tends to work near noon and midnight, but not so well near dawn and dusk. This puts an MLT bias in our data.

standing wave ratio (SWR). This tells about the amount of power in the standing wave versus the amount in the traveling wave. This is why our waves are not at exactly 90 degrees phase lag.